Lift AssociatedObjectStore to module top level; fold StateStore into State
`AssociatedObjectStore<Value>` is a generic primitive over Obj-C
associated objects with no knowledge of either tracking or
reward-verification. Having `RewardVerification` reach into the
`Tracking` namespace just to use it inverted the dependency between
two peer subsystems.

- Move the class out of `extension Tracking` into a top-level file
  `AssociatedObjectStore.swift`.
- The four tracking-specific delegate-store typealiases stay inside
  `Tracking`; renamed the file to `Tracking/DelegateStores.swift` to
  reflect what it now contains.
- `RewardVerification.StateStore.swift` (a 2-line typealias + singleton)
  is folded into `RewardVerification.State.swift` next to the State
  class it stashes; both files become one.

No behavior changes; 143/143 adapter tests still pass.
